add create meal option

+++ src/index.js
@@ -7,12 +7,13 @@ const addMealBtn = document.querySelector('#add-new-meal');
 const mealContainer = document.querySelector('#meal-container');
 let addMeal = false;
 const categoryAdapter = new CategoryAdapter(`${BASE_URL}/categories`)
+const mealAdapter = new MealAdapter(`${BASE_URL}/meals`);
 
 addMealBtn.addEventListener("click", () => {
     addMeal = !addMeal;
     if (addMeal) {
         categoriesContainer.style.display = "block";
-        submitNewMeal();
+        submitNewCategory();
     } else {
         categoriesContainer.style.display = "none";
     }
@@ -34,6 +35,7 @@ function createMenu(categoryObj) {
             <div class="category" data-category-id="category-${category.id}">
                 <a class='category-link' href="#">${category.attributes.name}</a>
                 <button class='delete-category-btn' data-category-delete-id='${category.id}'>x</button>
+                <button class='add-meal' data-category-meal-id='${category.id}'>Add Meal</button>
                 <ul class="meals">
                 </ul>
             </div>
@@ -41,16 +43,32 @@ function createMenu(categoryObj) {
     })
     const nodeListOfCategoryDivs = document.querySelectorAll('.category')
     nodeListOfCategoryDivs.forEach(function(div) {
-        let divButton = div.querySelector('.delete-category-btn')
-        divButton.addEventListener('click', function(e) {
-            let categoryId = e.target.dataset.categoryDeleteId;
-            categoryAdapter.deleteCategory(categoryId)
-                .then(function(categoryInfo){
-                    // console.log(categoryInfo.data.id)
-                    document.querySelector(`[data-category-id="category-${categoryInfo.data.id}"]`).remove();
-                    alert(`${categoryInfo.data.attributes.name} category has been deleted`);
-                })
-                .catch(errors => alert(errors));
+        // div.querySelectorAll('button').forEach(button => console.log(button))
+        // let deleteButton = div.querySelector('.delete-category-btn');
+        // let addButton = div.querySelector('.add-meal');
+        // console.log(addButton)
+        div.querySelectorAll('button').forEach(function(button) {
+            // console.log(button.className)
+            button.addEventListener('click', (e) => {
+                // console.log(e.target.dataset)
+                // THIS DELETES CATEGORY
+                if (e.target.className === "delete-category-btn") {
+                    let categoryId = e.target.dataset.categoryDeleteId;
+                    categoryAdapter.deleteCategory(categoryId)
+                        .then(function(categoryInfo){
+                            // console.log(categoryInfo.data.id)
+                            document.querySelector(`[data-category-id="category-${categoryInfo.data.id}"]`).remove();
+                            alert(`${categoryInfo.data.attributes.name} category has been deleted`);
+                        })
+                } else if (e.target.className === "add-meal") {
+                    const div = document.querySelector(`[data-category-id="category-${e.target.dataset.categoryMealId}"]`);
+                    const divId = e.target.dataset.categoryMealId;
+                    // console.log(div)
+                    div.innerHTML += renderMealForm();
+                    const mealForm = div.querySelector('#meal-form');
+                    submitNewMeal(mealForm, divId);
+                }
+            })
         })
     })
     categoryObj.included.forEach(function(meal)  {
@@ -61,6 +79,7 @@ function createMenu(categoryObj) {
             <li class='meal-title'>${meal.attributes.name}</li>
             <li><em>Description:</em> ${meal.attributes.description}</li>
             <li><em>Price:</em> $${meal.attributes.price}</li>
+            <li><button class='delete-meal' data-delete-meal-id='${meal.id}'>Delete Meal</button></li>
         `
         mealLink.addEventListener('click', (e) => {
             e.preventDefault();
@@ -72,11 +91,10 @@ function createMenu(categoryObj) {
             }
         })
     })
-    
 }
 
 function postCategory(data) {
-    configurationObject = {
+    let configurationObject = {
         method: "POST",
         headers: {
           "Content-Type": "application/json",
@@ -91,7 +109,8 @@ function postCategory(data) {
                 mealContainer.innerHTML += `
                     <div class="category" data-category-id="category-${obj.data.id}">
                         <a class='category-link' href="#">${obj.data.attributes.name}</a>
-                        <button class='delete-category-btn' data-category-delete-id='${obj.data.id}'>X</button>
+                        <button class='delete-category-btn' data-category-delete-id='${obj.data.id}'>x</button>
+                        <button class='add-meal' data-category-meal-id='${obj.data.id}'>Add Meal</button>
                         <ul class="meals">
                         </ul>
                     </div>
@@ -112,11 +131,69 @@ function formData() {
     };
 };
 
-function submitNewMeal() {
+function submitNewCategory() {
     categoryForm.addEventListener('submit', (e) => {
         e.preventDefault();
         // console.log(e.target)
         postCategory(formData());
     })
 }
- 
+
+function renderMealForm() {
+    return `
+        <form id='meal-form'>
+            <h3>Add Meal:</h3>
+            <label>Name:</label>
+            <input type='text' name='name' value='' placeholder="Cheeseburger, Steak, etc...">
+            <label>Description:</label>
+            <input type='text' name='description' value='' placeholder='Well Cook, lettuce, tomato, onion...'>
+            <label>Price: $</label>
+            <input type='number' name='price' value='' placeholder='14.95'>
+            <input type='submit' name='submit' value='Submit'>
+        </form>
+    `
+}
+
+function submitNewMeal(mealForm, divId) {
+    mealForm.addEventListener('submit', (e) => {
+        // console.log(e.target)
+        e.preventDefault();
+        postMeal(mealForm, mealFormData(mealForm, divId))
+    })
+}
+
+function mealFormData(mealForm, divId) {
+    return {
+        name: mealForm.name.value,
+        description: mealForm.description.value,
+        price: Number(mealForm.price.value).toFixed(2),
+        categoryId: divId
+    }
+}
+
+function postMeal(mealForm, data) {
+    let objectData = {
+        method: "POST",
+        headers: {
+          "Content-Type": "application/json",
+          "Accept": "application/json"
+        },
+        body: JSON.stringify(data)
+    };
+    mealAdapter.createNewMeal(objectData)
+        .then(function(json) {
+            // console.log(json)
+            if (!!json.data) {
+                mealForm.innerHTML += `
+                    <li class='meal-title'>${json.data.attributes.name}</li>
+                    <li><em>Description:</em> ${json.data.attributes.description}</li>
+                    <li><em>Price:</em> $${json.data.attributes.price}</li>
+                    <li><button class='delete-meal' data-delete-meal-id='${json.data.id}'>Delete Meal</button></li>
+                `
+                alert(`${json.data.attributes.name} was created successfully!`);
+                mealForm.reset();
+            } else {
+                alert(obj.message);
+            }
+        })
+}
